Fix tests regression
Delete SDC resources only if exists and archive all before delete
Bump onapsdk requirements

Issue-ID: INT-2193

+++ src/onaptests/steps/onboard/pnf.py
@@ -63,7 +63,9 @@ class PnfOnboardStep(BaseStep):
     @BaseStep.store_state(cleanup=True)
     def cleanup(self):
         pnf: Pnf = Pnf(name=settings.PNF_NAME)
-        pnf.delete()
+        if pnf.exists():
+            pnf.archive()
+            pnf.delete()
         super().cleanup()
 
 
@@ -136,5 +138,7 @@ class YamlTemplatePnfOnboardStep(YamlTemplateBaseStep):
         if "pnfs" in self.yaml_template:
             for pnf in self.yaml_template["pnfs"]:
                 pnf_obj: Pnf = Pnf(name=pnf["pnf_name"])
-                pnf_obj.delete()
+                if pnf_obj.exists():
+                    pnf_obj.archive()
+                    pnf_obj.delete()
         super().cleanup()
